Refactored the code so it plays nice.

+++ chicagothree/data.py
@@ -1,33 +1,12 @@
-import sqlalchemy
-from sqlalchemy import Column, Integer, String, ForeignKey, literal
-import sqlalchemy.orm 
-from sqlalchemy.orm import relationship, backref, Session
-import sqlalchemy.ext.declarative as declarative
-
-Base = declarative.declarative_base()
-
-class Tag(Base):
-    __tablename__='tbl_tag'
-    id = Column(Integer, primary_key=True)
-    name = Column(String(50), nullable=False, unique=True)
-    options = relationship('Option', backref='Tag')
-    
-    def __str__(self):
-        return '<Tag: %s %s>' % (self.name, self.default)
-
-class Option(Base):
-    __tablename__='tbl_option'
-    id = Column(Integer, primary_key=True)
-    tag_id = Column(Integer, ForeignKey('tbl_tag.id'))
-    value = Column(String(20), nullable=False)
-
-    def __str__(self):
-        return '<Option:  %s>' % (self.value)
+from sqlalchemy import literal
+from model import Base, Option, Tag
 
 def create_schema(session):
     Base.metadata.bind=session.bind
-    Base.metadata.drop_all()
+    if len(Base.metadata.tables)>0:
+        Base.metadata.drop_all()
     Base.metadata.create_all()
+    session.commit()
 
 def add_tag(tname, session):
     """
